// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const ProductStatusEnum = {
  AVAILABLE: 'AVAILABLE',
  OUT: 'OUT',
  LOW: 'LOW'
} as const

export type ProductStatusEnum = (typeof ProductStatusEnum)[keyof typeof ProductStatusEnum]


export const OrderStatusEnum = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatusEnum = (typeof OrderStatusEnum)[keyof typeof OrderStatusEnum]


export const SessionStatusEnum = {
  OPEN: 'OPEN',
  CLOSED: 'CLOSED'
} as const

export type SessionStatusEnum = (typeof SessionStatusEnum)[keyof typeof SessionStatusEnum]


export const PaymentMethodEnum = {
  CASH: 'CASH',
  POS_CARD: 'POS_CARD',
  BANK_TRANSFER: 'BANK_TRANSFER',
  STORE_CREDIT: 'STORE_CREDIT',
  SPLIT: 'SPLIT'
} as const

export type PaymentMethodEnum = (typeof PaymentMethodEnum)[keyof typeof PaymentMethodEnum]


export const PaymentStatusEnum = {
  PAID: 'PAID',
  PARTIALLY_PAID: 'PARTIALLY_PAID',
  PENDING: 'PENDING'
} as const

export type PaymentStatusEnum = (typeof PaymentStatusEnum)[keyof typeof PaymentStatusEnum]


export const StockMovementType = {
  PURCHASE: 'PURCHASE',
  SALE: 'SALE',
  RETURN: 'RETURN',
  ADJUSTMENT: 'ADJUSTMENT',
  BREAK_BULK: 'BREAK_BULK'
} as const

export type StockMovementType = (typeof StockMovementType)[keyof typeof StockMovementType]


export const UserRole = {
  ADMIN: 'ADMIN',
  STORE_MANAGER: 'STORE_MANAGER',
  CASHIER: 'CASHIER',
  SALES_REP: 'SALES_REP',
  INVENTORY_CONTROLLER: 'INVENTORY_CONTROLLER',
  ACCOUNTANT: 'ACCOUNTANT',
  SUPPLIER_MANAGER: 'SUPPLIER_MANAGER',
  STAFF: 'STAFF',
  MANAGER: 'MANAGER',
  SUPPORT: 'SUPPORT'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatusEnum = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  SUSPENDED: 'SUSPENDED'
} as const

export type UserStatusEnum = (typeof UserStatusEnum)[keyof typeof UserStatusEnum]